tests/test_mobile.py
# Placeholder for Mobile tests
# Directory: tests/test_mobile.py
from appium import webdriver as appium_webdriver
import pytest

@pytest.mark.parametrize("platform_name, platform_version, device_name", [
    ("Android", "11.0", "emulator-5554"),
    ("iOS", "15.0", "iPhone 12")
])
def test_mobile_app(platform_name, platform_version, device_name):
    desired_caps = {
        "platformName": platform_name,
        "platformVersion": platform_version,
        "deviceName": device_name,
        "app": "path/to/app.apk" if platform_name == "Android" else "path/to/app.ipa"
    }

    driver = appium_webdriver.Remote("http://localhost:4723/wd/hub", desired_caps)
    # Example test: Verify app launches
    assert driver.is_app_installed(desired_caps['app'])
    driver.quit()
