requirements.txt
selenium
pytest
pytest-allure-adaptor
requests
sqlalchemy
appium-python-client
pywinauto